Move: 新たなmigrationのモデルをModelsフォルダに移動
